+++ src/modules/profile/pages/profilePage/views/account/views/badges/BadgeItem.tsx
@@ -1,49 +1,18 @@
 import { Image, VStack } from '@chakra-ui/react'
 
-// import { useState } from 'react'
-// import { useTranslation } from 'react-i18next'
-// import { ButtonComponent } from '../../../../../../../../components/ui'
 import { UserBadge } from '../../../../../../../../types'
 
-interface BadgeItemProps {
+type BadgeItemProps = {
   userBadge: UserBadge
   handleClick: (badge: UserBadge) => void
 }
 
-export const BadgeItem = ({
-  userBadge,
-  handleClick,
-}: BadgeItemProps) => {
-  // const { t } = useTranslation()
+export const BadgeItem = ({ userBadge, handleClick }: BadgeItemProps) => {
   const { badge } = userBadge
 
-  // const [claiming, setClaiming] = useState(false)
-  // const onClaim = (e: React.MouseEvent<HTMLButtonElement>) => {
-  //   e.preventDefault()
-  //   e.stopPropagation()
-  //   claimABadge({
-  //     userBadgeId: userBadge.id,
-  //     badgeId: badge.uniqueName,
-  //     badgeAwardId: userBadge.badgeAwardEventId || '',
-  //     isClaiming: setClaiming,
-  //   })
-  // }
-
   return (
-    <VStack
-      key={userBadge.id}
-      overflow="hidden"
-      w="full"
-      onClick={() => handleClick(userBadge)}
-      _hover={{ cursor: 'pointer' }}
-    >
+    <VStack overflow="hidden" w="full" onClick={() => handleClick(userBadge)} _hover={{ cursor: 'pointer' }}>
       <Image width="auto" maxWidth="110px" src={badge.image} alt={`${badge.name} badge image`} />
-      {/* {
-        !isClaimed && isEdit &&
-        <ButtonComponent size="sm" primary onClick={onClaim} isLoading={claiming}>
-          {t('Claim')}
-        </ButtonComponent>
-      } */}
     </VStack>
   )
 }