+++ src/shared/constants/platform/wallet.ts
@@ -1,5 +1,1 @@
-export enum WalletConnectDetails {
-  LightningAddressConnectionDetails = 'LightningAddressConnectionDetails',
-}
-
 export const LIGHTNING_FEE_PERCENTAGE = 5